replace videoEnhance.js with replaceControls within video.js
